% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eros_fish.R
\name{calc_eros_fish}
\alias{calc_eros_fish}
\title{Formats data and calculates values of biomass, density, and
bioerosion rates}
\usage{
calc_eros_fish(data, rates_dbase = c("Perry", "Kindinger"))
}
\arguments{
\item{data}{csv file of imported data to be formatted and for biomass and
bioerosion to be calculated from the COUNT and sizebin input data}

\item{rates_dbase}{Use Perry or Kindinger rates}
}
\description{
Formats data and calculates values of biomass, density, and
bioerosion rates
}
\author{
Rebecca Weible
}
