% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_belt.R
\name{calc_fish_belt}
\alias{calc_fish_belt}
\title{Calculate parrotfish erosion rates from belt data}
\usage{
calc_fish_belt(data, dbasetype = c("IPRB", "Kindinger"), full_summary = TRUE)
}
\arguments{
\item{data}{fish survey data.}

\item{dbasetype}{Erosion rates database to use. Choose either Indo-Pacific
ReefBudget ("dbase_type = "IPRB") or U.S. Pacific Islands rates developed
by Tye Kindinger, NOAA PIFSC ("dbase_type = "Kindinger").}
}
\description{
Calculate parrotfish erosion rates from belt data
}
\examples{
fish_data <- read.csv("CB_FishBelt_alldata.csv", na = "", check.names = FALSE)

fish_belt <- calc_fish_belt(data = fish_data, dbase_type = "Kindinger",
full_summary = TRUE)
}
\author{
Rebecca Weible
}
