% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_distance_matrix.R
\name{calc_fish_distance_matrix}
\alias{calc_fish_distance_matrix}
\title{Calculate the shortest distance between each fixed SPC/OCC site and each fish REA site 'as a fish swims'}
\usage{
calc_fish_distance_matrix(
  pointsOCC,
  pointsFISH,
  island_poly,
  resolution_m = 50
)
}
\arguments{
\item{pointsOCC}{Lat/Long of fixed OCC site, or fixed SPC data, aka target sites.}

\item{pointsFISH}{Lat/Long of fish REA sites, or stratified random sampled SPC surveys, aka all other survey sites.}

\item{island_poly}{Simple Feature polygon Shapefile of Pacific Islands where surveys were conducted.}

\item{resolution_m}{Resolution of the output raster plot displaying distance from target point to survey point of interest.}
}
\description{
Calculate the shortest distance between each fixed SPC/OCC site and each fish REA site 'as a fish swims'
}
\examples{
distance_matrix <- calc_fish_distance_matrix(pointsOCC = ptsOCC_ISL, pointsFISH = ptsFISH_ISL, island_poly = poly_ISL, resolution_m = 50)
}
\author{
Thomas Oliver

Rebecca Weible
}
