% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_fixed_spc.R
\name{calc_fish_fixed_spc}
\alias{calc_fish_fixed_spc}
\title{Calculate parrotfish erosion rates from fixed spc data}
\usage{
calc_fish_fixed_spc(data, dbase_type = c("IPRB", "Kindinger"))
}
\arguments{
\item{data}{fish survey data.}

\item{dbase_type}{Erosion rates database to use. Choose either Indo-Pacific
ReefBudget ("dbase_type = "IPRB") or U.S. Pacific Islands rates developed
by Tye Kindinger, NOAA PIFSC ("dbase_type = "Kindinger").}
}
\description{
Calculate parrotfish erosion rates from fixed spc data
}
\examples{
fish_data <- read.csv("CB_FishBelt_alldata.csv", na = "", check.names = FALSE)

fish_fixed_spc <- calc_fish_fixed_spc(data = fish_data, dbase_type = "Kindinger")
}
\author{
Rebecca Weible
}
