% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_metrics.R
\name{calc_fish_metrics}
\alias{calc_fish_metrics}
\title{Calculate bite rate, volume, and proportion of scars from bioerosion equations}
\usage{
calc_fish_metrics(
  bio_equations,
  tl2fl_conversion,
  avg_sst,
  avg_time,
  avg_depth
)
}
\arguments{
\item{bio_equations}{dataframe with Tye's equations of bite rate,
bite volume, and prop of scars}

\item{tl2fl_conversion}{dataframe with Tyes TL to FL conversion constants}

\item{avg_sst}{Input average Sea Surface Temperature (Celsius) of
location where data was collected (HI, MARIAN, AMSM, PRIAS, etc.)}

\item{avg_time}{Input average time period (HHMM) of location where data were
collected (e.g. if collected between 8am and 1pm, average will be 11am)}

\item{avg_depth}{Input average depth (Meters) at location where data were
collected}
}
\description{
Calculate bite rate, volume, and proportion of scars from bioerosion equations
}
\author{
Rebecca Weible
}
