% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_strs_spc.R
\name{calc_fish_strs_spc}
\alias{calc_fish_strs_spc}
\title{Calculate parrotfish erosion rates from associated spc data}
\usage{
calc_fish_strs_spc(data, rates_dbase = rates_dbase)
}
\arguments{
\item{data}{fish survey data.}

\item{rates_dbase}{Erosion rates database to use.}
}
\description{
Calculate parrotfish erosion rates from associated spc data
}
\examples{
fish_data <- read.csv("CB_FishBelt_alldata.csv", na = "", check.names = FALSE)

fish_strs_spc <- calc_fish_strs_spc(data = fish_data, rates_dbase = rates_dbase)
}
\author{
Rebecca Weible
}
