% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kindinger_erosion_metrics.R
\name{calc_kindinger_erosion_metrics}
\alias{calc_kindinger_erosion_metrics}
\title{Calculate Kindinger erosion rate metrics: bite rate, volume, and proportion of scars}
\usage{
calc_kindinger_erosion_metrics(
  kindinger_equations_csv,
  kindinger_functional_groups_csv,
  sister_species_csv,
  SST_Input_Celsius
)
}
\arguments{
\item{kindinger_equations_csv}{.csv file in Kindinger et al AppendixS2 final format,
downloaded as is.}

\item{kindinger_functional_groups_csv}{.csv file in NCRMP herbivore functional groups format,
downloaded as is.}

\item{sister_species_csv}{.csv file with parrot fish sister species assignments,
downloaded as is.}

\item{SST_Input_Celsius}{average sea surface temperature in celsius from region data was
collected.

#'@import tidyverse}
}
\description{
Calculate Kindinger erosion rate metrics: bite rate, volume, and proportion of scars
}
\author{
Rebecca Weible
}
