% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prod.R
\name{calc_prod}
\alias{calc_prod}
\title{Calculate production rates for each benthic component measured on
a benthic carbonate budgets transect}
\usage{
calc_prod(
  substrate_class,
  substrate_code,
  morphology_code,
  substrate_cover_cm,
  region_code,
  prod_dbase
)
}
\arguments{
\item{substrate_class}{Type of substrate.}

\item{substrate_code}{NCRMP taxa code}

\item{morphology_code}{Taxa morphology (corals only)}

\item{substrate_cover_cm}{Surface distance of benthic component}

\item{region_code}{Survey region}

\item{prod_dbase}{Production database to reference}
}
\description{
Calculate production rates for each benthic component measured on
a benthic carbonate budgets transect
}
\examples{
calc_prod("CORAL","PLOB","MD",10,"MHI",prod_dbase)


}
\author{
Hannah Barkley
}
