% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sfm.r
\name{combine_sfm}
\alias{combine_sfm}
\title{Combine individual SfM transect files}
\usage{
combine_sfm(
  sfm_folder,
  region_code,
  location_code,
  local_date,
  cruise_id,
  surveyor
)
}
\arguments{
\item{sfm_folder}{Directory containing individual SfM transect files}

\item{region_code}{Region code where surveys were conducted (e.g. "MHI")}

\item{location_code}{Location code where surveys were conducted (e.g. "OAH")}

\item{local_date}{Local Date of imagery collection (yyyy-mm-dd)}

\item{cruise_id}{Cruise ID for survey collection}

\item{surveyor}{Initials of data collector}
}
\description{
Combine individual SfM transect files
}
\author{
Hannah Barkley
}
