% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_iprb.r
\name{convert_iprb}
\alias{convert_iprb}
\title{Convert all individual Indo-Pacific ReefBudget (IPRB) Excel
Entry sheets for each site into one large data frame}
\usage{
convert_iprb(iprb_csv, iprb_metadata, add_location, occ_siteid)
}
\arguments{
\item{iprb_csv}{.csv file in Indo-Pacific ReefBudget Excel sheet format,
downloaded as is.}

\item{iprb_metadata}{.csv file with all of the metadata from
Indo-Pacific ReefBudget Excel sheet, saved in a dataframe format.}

\item{add_location}{create column for unique, general site name
where the data comes from ("Kaneohe Bay", "Tumon Bay", ...).}

\item{occ_siteid}{OCC site ID ("OCC-OAH-005", "OCC-GUA_015", ...).}
}
\description{
Convert all individual Indo-Pacific ReefBudget (IPRB) Excel
Entry sheets for each site into one large data frame
}
\author{
Rebecca Weible
}
