% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fish_assoc_sites.R
\name{create_fish_assoc_sites}
\alias{create_fish_assoc_sites}
\title{Create fish associated SPC sites in relation to fixed sites}
\usage{
create_fish_assoc_sites(data, fixed_metadata, subset_distance_m)
}
\arguments{
\item{data}{fish survey data.}

\item{subset_distance_m}{Assigned associated site distances, in meters, from fixed SPC/OCC site to all other fish SPC sites, based on parrotfish foraging boundaries.}
}
\description{
Create fish associated SPC sites in relation to fixed sites
}
\examples{
fish_data <- read.csv("ESD_CarbBudget_SPC_OAHU_2021.csv", na = "", check.names = FALSE)

sites_associated_dbase <- create_fish_assoc_sites(data = fish_data, subset_distance_m = 2000)
}
\author{
Thomas Oliver

Kisei Tanaka

Rebecca Weible
}
