% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_4ncei.R
\name{format_4ncei}
\alias{format_4ncei}
\title{Format net production data for NCEI submission}
\usage{
format_4ncei(data)
}
\arguments{
\item{data}{output of \code{process_net}.}
}
\description{
Format net production data for NCEI submission
}
\author{
Hannah Barkley
}
