% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_benthic_sfm.R
\name{format_benthic_sfm}
\alias{format_benthic_sfm}
\title{Combine and format individual SfM benthic data transect files}
\usage{
format_benthic_sfm(site_file, sfm_folder)
}
\arguments{
\item{site_file}{CSV file with site info.}

\item{sfm_folder}{Directory containing SfM transect files.}
}
\description{
Combine and format individual SfM benthic data transect files
}
\author{
Hannah Barkley
}
