% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_iprb_erosion_rates.R
\name{format_iprb_erosion_rates}
\alias{format_iprb_erosion_rates}
\title{Format Indo-Pacific ReefBudget parrotfish erosion rates}
\usage{
format_iprb_erosion_rates(iprb_rates, output = c("all", "finalerosion"))
}
\arguments{
\item{iprb_rates}{Indo-Pacific ReefBudget parrotfish erosion rates
("Equations" tab in IP_Parrotfish_erosion_rates_database_v1.3.xlsx,
downloaded from https://geography.exeter.ac.uk/reefbudget/indopacific/).}
}
\description{
Format Indo-Pacific ReefBudget parrotfish erosion rates
}
\author{
Rebecca Weible
}
