% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_rates.R
\name{format_rates}
\alias{format_rates}
\title{Format input data}
\usage{
format_rates(iprb_rates, output = c("all", "finalerosion"))
}
\arguments{
\item{iprb_rates}{.csv file of Chris Perry's EXACT Equations tab}
}
\description{
Format input data
}
\author{
Rebecca Weible
}
