% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sfm.R
\name{format_sfm}
\alias{format_sfm}
\title{Format SfM production data for processing}
\usage{
format_sfm(data)
}
\arguments{
\item{data}{SfM data set; output of \code{combine_sfm}}
}
\description{
Format SfM production data for processing
}
\author{
Hannah Barkley
}
