% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sfm.R
\name{prep_sfm}
\alias{prep_sfm}
\title{Prepare SfM production data for processing}
\usage{
prep_sfm(sfm_folder, sites_metadata)
}
\arguments{
\item{sfm_folder}{Directory containing SfM transect files.}

\item{sites_metadata}{Data frame containing information about sites}
}
\description{
Prepare SfM production data for processing
}
\author{
Hannah Barkley
}
