% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_spc.R
\name{prep_spc}
\alias{prep_spc}
\title{Reformat NCRMP stratified random SPC data}
\usage{
prep_spc(data_spc, fixed_metadata = NULL)
}
\arguments{
\item{data_spc}{Fish spc survey data.}

\item{fixed_metadata}{Data frame with fixed site metadata}
}
\description{
Reformat NCRMP stratified random SPC data
}
\author{
Hannah Barkley
}
