% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_urchins.R
\name{prep_urchins}
\alias{prep_urchins}
\title{Prep urchin census data}
\usage{
prep_urchins(data, sites_metadata)
}
\arguments{
\item{data}{Urchin observation data set.}

\item{sites_metadata}{Data frame containing information about sites}
}
\description{
Prep urchin census data
}
\author{
Hannah Barkley
}
