% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_fish.R
\name{process_fish}
\alias{process_fish}
\title{Process parrotfish erosion rates from belt and spc data}
\usage{
process_fish(
  spc_data = NULL,
  belt_data = NULL,
  method_type = c("Fixed", "StRS", "Mean StRS", "Belt"),
  fixed_metadata = NULL,
  dbase_type = c("Kindinger", "IPRB"),
  subset_distance_m = 6000
)
}
\arguments{
\item{spc_data}{Fish SPC survey data (default is NULL).}

\item{belt_data}{Fish Belt survey data (default is NULL).}

\item{method_type}{Type of survey data to process. "Fixed" (fixed SPC only), "StRS" (stratified random SPC), "Mean StRS" (average of nearby stratified random SPC), or "Belt" (belt data only)}

\item{fixed_metadata}{Dataframe with fixed site info}

\item{dbase_type}{Erosion rates database to use. Choose either Indo-Pacific
ReefBudget ("dbase_type = "IPRB") or U.S. Pacific Islands rates developed
by Tye Kindinger, NOAA PIFSC ("dbase_type = "Kindinger").}

\item{subset_distance_m}{Assigned associated site distances, in meters, from fixed SPC/OCC site to all other fish SPC sites, based on parrotfish foraging boundaries.}
}
\description{
Process parrotfish erosion rates from belt and spc data
}
\examples{
fish_data_belt <- read.csv("CB_FishBelt_alldata.csv", na = "", check.names = FALSE)

fish_belt <- process_fish(spc_data = fish_data_spc,
dbase_type = "Kindinger", belt_data = fish_data_belt, subset_distance_m = 2000)
}
\author{
Rebecca Weible
}
