% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_net.R
\name{process_net}
\alias{process_net}
\title{Calculate net production rates for each transect}
\usage{
process_net(prod, urch, fish, format = "wide")
}
\arguments{
\item{prod}{Production data}

\item{urch}{Urchin data}

\item{parrotfish}{Parrotfish data}
}
\description{
Calculate net production rates for each transect
}
\author{
Hannah Barkley
}
