% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prod.R
\name{process_prod}
\alias{process_prod}
\title{Calculate carbonate production rates from benthic census data}
\usage{
process_prod(
  data,
  transect_id = c("A1", "A2", "A3", "B1", "B2", "B3"),
  transect_length = c(10, 10, 10, 10, 10, 10),
  dbase_type = c("IPRB", "NCRMP"),
  method_name = c("IPRB", "Chords"),
  data_type = c("In water", "SfM"),
  full_summary = FALSE,
  label = NULL,
  ...
)
}
\arguments{
\item{data}{Benthic field data set}

\item{transect_id}{String of transect names}

\item{transect_length}{String of transect lengths in meters}

\item{dbase_type}{Production database to use ("IPRB" or "NCRMP")}

\item{method_name}{Transect design ("IPRB" or "Chords")}

\item{data_type}{Type of data collection ("In water" or "SfM")}
}
\description{
Calculate carbonate production rates from benthic census data
}
\author{
Hannah Barkley
}
