% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_urchins.R
\name{process_urchins}
\alias{process_urchins}
\title{Calculate urchin erosion rates from urchin census data}
\usage{
process_urchins(data, full_summary = TRUE)
}
\arguments{
\item{data}{Urchin observation data set.}

\item{full_summary}{Logical. If TRUE, returns a list of dataframes.}
}
\description{
Calculate urchin erosion rates from urchin census data
}
\author{
Hannah Barkley
}
