% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_calc_prod.R
\name{run_calc_prod}
\alias{run_calc_prod}
\title{Run calc_prod function}
\usage{
run_calc_prod(data, transect_id, transect_length, dbase_type, data_type, ...)
}
\arguments{
\item{data}{Benthic data set to process}

\item{transect_id}{String of transect names (e.g., ("A1", "A2", "A3", "B1", "B2", "B3")).}

\item{transect_length}{String of transect lengths in meters (e.g., c(10, 10, 10, 10, 10, 10)).}

\item{dbase_type}{Production database to use, either Indo-Pacific ReefBudget ("IPRB")
or U.S. Pacific Islands NCRMP-specific database ("NCRMP"). The Indo-Pacific ReefBudget
database is derived from "IP Calcification and bioerosion rates database v.1.3",
downloaded from https://geography.exeter.ac.uk/reefbudget/indopacific/.}

\item{data_type}{Type of data collection ("In water" or "SfM").}
}
\description{
Run calc_prod function
}
\author{
Hannah Barkley
}
