% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fish_erosion.R
\name{summarize_fish_erosion}
\alias{summarize_fish_erosion}
\title{Summarize parrotfish bioerosion rates per site}
\usage{
summarize_fish_erosion(species_table, full_summary = FALSE)
}
\arguments{
\item{species_table}{Species table with average metrics}

\item{full_summary}{Return full summary of results (default is FALSE)}
}
\description{
Summarize parrotfish bioerosion rates per site
}
\author{
Rebecca Weible
}
