% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fish_metrics.R
\name{summarize_fish_metrics}
\alias{summarize_fish_metrics}
\title{Calculate metrics for density, biomass, and bioerosion at the transect
and site level}
\usage{
summarize_fish_metrics(
  data,
  metric = c("density", "biomass", "bioerosion"),
  level = c("transect", "site"),
  summarize_by = c("size class", "species", "overall")
)
}
\arguments{
\item{data}{output from \code{calc_eros_fish_output}.}

\item{metric}{Metric to summarize ("count", "biomass", or "bioerosion").}

\item{level}{Level to summarize at ("transect" or "site" level).}

\item{summarize_by}{Grouping factor to summarize by ("size class",
"species", or "overall").}
}
\description{
Calculate metrics for density, biomass, and bioerosion at the transect
and site level
}
\examples{
calc_eros_fish_output <- calc_eros_fish(data, rates_dbase = rates_dbase)
}
\author{
Rebecca Weible
}
