% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_prod.R
\name{summarize_prod}
\alias{summarize_prod}
\title{Summarize carbonate production rates at transect and site level}
\usage{
summarize_prod(
  data,
  transect_summary,
  dbase_type = c("IPRB", "NCRMP"),
  summarize_by = c("substrate code", "substrate class", "coral group", "overall"),
  level = c("transect", "site"),
  macro_rate = 0.209,
  macro_rate_ci = 0.129,
  micro_rate = 0.262,
  micro_rate_ci = 0.18,
  ...
)
}
\arguments{
\item{data}{Data set to summarize; product of run_calc_prod}

\item{transect_summary}{Transect sumamry; product of run_calc_prod}

\item{dbase_type}{Production database to use ("IPRB" or "NCRMP")}

\item{summarize_by}{Grouping factor to summarize by ("substrate code",
"substrate class", "coral group",or "overall")}

\item{level}{Summarize at "transect" or "site" level}

\item{macro_rate}{Rate of macrobioerosion}

\item{macro_rate_ci}{Confidence interval for rate of macrobioerosion}

\item{micro_rate}{Rate of microbioerosion}

\item{micro_rate_ci}{Confidence interval for rate of microbioerosion}
}
\description{
Summarize carbonate production rates at transect and site level
}
\author{
Hannah Barkley
}
