% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_transect.R
\name{summarize_transect}
\alias{summarize_transect}
\title{Summarize benthic transect metadata}
\usage{
summarize_transect(
  data,
  transect_id,
  transect_length,
  method_name = c("IPRB", "Chords", "SfM")
)
}
\arguments{
\item{data}{Production data set.}

\item{transect_id}{String of transect names (e.g., ("A1", "A2", "A3", "B1", "B2", "B3")).}

\item{transect_length}{String of transect lengths in meters (e.g., c(10, 10, 10, 10, 10, 10)).}

\item{method_name}{Method by which benthic data were collected ("IPRB" "Chords", or "SfM").}
}
\description{
Summarize benthic transect metadata
}
\examples{
summarize_transect(
    data = data,
    transect_id = c("A1", "A2", "A3", "B1", "B2", "B3")
    transect_length = c(10, 10, 10, 10, 10, 10),
    method_name = "IPRB"
)

}
\author{
Hannah Barkley
}
