% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_belt.R
\name{calc_fish_belt}
\alias{calc_fish_belt}
\title{Calculate parrotfish erosion rates from belt data}
\usage{
calc_fish_belt(data, rates_dbase = rates_dbase, full_summary = TRUE)
}
\arguments{
\item{data}{fish survey data.}

\item{rates_dbase}{Erosion rates database to use.}
}
\description{
Calculate parrotfish erosion rates from belt data
}
\examples{
fish_data <- read.csv("CB_FishBelt_alldata.csv", na = "", check.names = FALSE)

fish_belt <- calc_fish_belt(data = fish_data, dbase_type = "Kindinger",
full_summary = TRUE)
}
\author{
Rebecca Weible
}
