% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_fixed_spc.R
\name{calc_fish_fixed_spc}
\alias{calc_fish_fixed_spc}
\title{Calculate parrotfish erosion rates from fixed spc data}
\usage{
calc_fish_fixed_spc(
  data,
  rates_dbase = rates_dbase,
  shape_file = fish_pacific_islands_shapefile
)
}
\arguments{
\item{data}{fish survey data.}

\item{rates_dbase}{Erosion rates database to use.}

\item{shape_file}{shapefiles of pacific islands.}
}
\description{
Calculate parrotfish erosion rates from fixed spc data
}
\examples{
fish_data <- read.csv("CB_FishBelt_alldata.csv", na = "", check.names = FALSE)

fish_fixed_spc <- calc_fish_fixed_spc(data = fish_data, dbasetype = rates_dbase, shape_file = fish_pacific_islands_shapefile)
}
\author{
Rebecca Weible
}
