% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_metrics.R
\name{calc_fish_metrics}
\alias{calc_fish_metrics}
\title{Calculate bite rate, volume, and proportion of scars from bioerosion equations}
\usage{
calc_fish_metrics(
  bio_equations,
  tl2fl_conversion,
  avg_sst,
  avg_time,
  avg_depth
)
}
\arguments{
\item{bio_equations}{Data frame with equations of bite rate,
bite volume, and prop of scars.}

\item{tl2fl_conversion}{Data frame with total length (TL) to fork length (FL)
conversion constants.}

\item{avg_sst}{Average Sea Surface Temperature (SST, in degrees Celsius) of
location where data were collected.}

\item{avg_time}{Average time period (HHMM) of location where data were
collected (e.g. if collected between 0800 and 1300, average will be 1100).}

\item{avg_depth}{Average depth (in meters) at location where data were
collected.}
}
\description{
Calculate bite rate, volume, and proportion of scars from bioerosion equations
}
\author{
Rebecca Weible
}
