% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_strs_spc.R
\name{calc_fish_strs_spc}
\alias{calc_fish_strs_spc}
\title{Calculate parrotfish erosion rates from associated spc data}
\usage{
calc_fish_strs_spc(
  data,
  dbase_type = c("IPRB", "Kindinger"),
  sites_associated = c("OAH", "MARIAN")
)
}
\arguments{
\item{data}{fish survey data.}

\item{dbase_type}{Erosion rates database to use. Choose either Indo-Pacific
ReefBudget ("dbase_type = "IPRB") or U.S. Pacific Islands rates developed
by Tye Kindinger, NOAA PIFSC ("dbase_type = "Kindinger").
#'@param sites_associated Location data was collected. Choose either Oahu ("sites_associated = "OAH"),
or Mariana Islands ("sites_associated = "MARIAN").}
}
\description{
Calculate parrotfish erosion rates from associated spc data
}
\examples{
fish_data <- read.csv("CB_FishBelt_alldata.csv", na = "", check.names = FALSE)

fish_belt <- process_fish(data = fish_data, dbase_type = "Kindinger",
sites_associated = "OAH")
}
\author{
Rebecca Weible
}
