% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prod.R
\name{calc_prod}
\alias{calc_prod}
\title{Calculate production rates for each benthic component measured on
a carbonate budget transect}
\usage{
calc_prod(
  substrate_class,
  substrate_code,
  morphology_code,
  substrate_cover_cm,
  prod_dbase
)
}
\arguments{
\item{substrate_class}{Type of substrate observed ("CORAL", "CCA", "TURF", "MA", ...).}

\item{substrate_code}{Taxa code observed (e.g. "PLOB", "MCAP", "PMEA", "CCA", ...).}

\item{morphology_code}{Taxa morphology observed (corals only; "BR", "MD", "EM", ...). Non-corals default to NA.}

\item{substrate_cover_cm}{Measure surface distance of benthic component, in cm.}

\item{prod_dbase}{Production database to reference, either Indo-Pacific ReefBudget ("IPRB")
or NCRMP-specific ("NCRMP"). Defaults to "NCRMP".}

\item{region_code}{Survey region ("MHI", "MARIAN", ...).}
}
\description{
Calculate production rates for each benthic component measured on
a carbonate budget transect
}
\details{
See included carbonate production databases \code{prod_dbase_iprb} and \code{prod_dbase_ncrmp}
for acceptable \code{substrate_class}, \code{substrate_code}, and \code{morphology_code} values.
}
\examples{
calc_prod(
substrate_class = "CORAL",
substrate_code = "PLOB",
morphology_code = "MD",
substrate_cover_cm = 10,
region_code = "MHI",
prod_dbase = "NCRMP")


}
\author{
Hannah Barkley
}
