% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sfm.r
\name{combine_sfm}
\alias{combine_sfm}
\title{Combine individual SfM benthic data transect files}
\usage{
combine_sfm(sfm_folder, region_code, cruise_id)
}
\arguments{
\item{sfm_folder}{Directory containing SfM transect files.}

\item{cruise_id}{Cruise ID for survey collection ("MP2108", "RA2201", ...).}

\item{region_codeSurvey}{region ("MHI", "MARIAN", ...).}
}
\description{
Combine individual SfM benthic data transect files
}
\examples{
combine_sfm(
   sfm_folder = "...",
   region_code = "MHI",
   local_date = "2020-01-01",
   cruise_id = "MP2108",
   surveyor = "HCB"
)
}
\author{
Hannah Barkley
}
