% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_iprb.r
\name{convert_iprb}
\alias{convert_iprb}
\title{Convert all individually saved Indo-Pacific ReefBudget (IPRB) Excel Entry sheets for each site into one large dataframe}
\usage{
convert_iprb(iprb_csv, iprb_metadata, add_location, OCC_SITEID)
}
\arguments{
\item{iprb_csv}{.csv file in Chris Perry's Excel sheet format without any changes}

\item{METADATA}{.csv file with all of the metadata from Chris Perry's Excel sheet saved in a dataframe format}

\item{ADDLOCATION}{create column for unique, general site name where the data comes from (i.e. "Kaneohe)}

\item{OCC.SITEID}{create column for unique OCC site ID (i.e. OCC-OAH-005)}
}
\description{
Convert all individually saved Indo-Pacific ReefBudget (IPRB) Excel Entry sheets for each site into one large dataframe
}
\author{
Rebecca Weible
}
