% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fish_assoc_sites.R
\name{create_fish_assoc_sites}
\alias{create_fish_assoc_sites}
\title{Create fish associated SPC sites in relation to fixed sites}
\usage{
create_fish_assoc_sites(data)
}
\arguments{
\item{data}{fish survey data.}
}
\description{
Create fish associated SPC sites in relation to fixed sites
}
\examples{
fish_data <- read.csv("CB_FishBelt_alldata.csv", na = "", check.names = FALSE)

sites_associated_dbase <- create_fish_assoc_sites(data = fish_data)
}
\author{
Rebecca Weible
}
