% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fish_spc.R
\name{format_fish_spc}
\alias{format_fish_spc}
\title{Format Stationary Point Count data}
\usage{
format_fish_spc(data, method = c("CbB", "nSPC"), rates_dbase_ = "IPRB")
}
\arguments{
\item{data}{all stationary point count data.}

\item{method}{type of SPC survey conducted. Choose either Fixed site SPC
("method = "CbB") or stratified random sampling SPC ("method = "nSPC").}

\item{rates_dbase_}{Erosion rates database to use. Choose either Indo-Pacific
ReefBudget ("rates_dbase_ = "IPRB") or U.S. Pacific Islands rates developed
by Tye Kindinger, NOAA PIFSC ("rates_dbase_ = "Kindinger").}
}
\description{
Format Stationary Point Count data
}
\author{
Rebecca Weible
}
