% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fish_spc.R
\name{format_fish_spc}
\alias{format_fish_spc}
\title{Format Stationary Point Count data}
\usage{
format_fish_spc(data, method = c("IPRB", "nSPC"), rates_dbase)
}
\arguments{
\item{data}{all stationary point count data.}

\item{method}{type of SPC survey conducted. Choose either Fixed site SPC
("method = "CbB") or stratified random sampling SPC ("method = "nSPC").}

\item{rates_dbase}{Erosion rates database to use.}
}
\description{
Format Stationary Point Count data
}
\author{
Rebecca Weible
}
