% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_fish.R
\name{process_fish}
\alias{process_fish}
\title{Calculate erosion rates from parrotfish belt data}
\usage{
process_fish(data, rates_dbase = "Kindinger", full_summary = FALSE)
}
\arguments{
\item{data}{Parrotfish belt data}

\item{rates_dbase}{Use Perry or Kindinger rates}
}
\description{
Calculate erosion rates from parrotfish belt data
}
\author{
Rebecca Weible
}
