% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_fish.R
\name{process_fish}
\alias{process_fish}
\title{Calculate parrotfish erosion rates from belt and spc data}
\usage{
process_fish(
  data,
  method = c("IPRB", "Fixed SPC", "StRS SPC"),
  dbase_type = c("IPRB", "Kindinger"),
  sites_associated = c("OAH", "MARIAN"),
  full_summary = TRUE
)
}
\arguments{
\item{data}{fish survey data.}

\item{method}{survey method. Choose either Fixed Belt ("method = "IPRB"),
Fixed Stationary Point Count ("method = "Fixed SPC"), or Associated Stationary
Point Count ("method = "StRS SPC").}

\item{dbase_type}{Erosion rates database to use. Choose either Indo-Pacific
ReefBudget ("rates_dbase_ = "IPRB") or U.S. Pacific Islands rates developed
by Tye Kindinger, NOAA PIFSC ("rates_dbase_ = "Kindinger").
#'@param sites_associated Location data was collected. Choose either Oahu ("method = "OAH"),
or Mariana Islands ("method = "MARIAN").}
}
\description{
Calculate parrotfish erosion rates from belt and spc data
}
\examples{
fish_data <- read.csv("CB_FishBelt_alldata.csv", na = "", check.names = FALSE)

fish_belt <- process_fish(data = fish_data, method = "Fixed SPC", rates_dbase = "Kindinger",
full_summary = TRUE)
}
\author{
Rebecca Weible
}
