% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prod.R
\name{process_prod}
\alias{process_prod}
\title{Calculate carbonate production rates from benthic census data}
\usage{
process_prod(
  data,
  dbase_type = "NCRMP",
  full_summary = TRUE,
  label = NULL,
  qc_check = FALSE,
  sites_metadata = NULL,
  ...
)
}
\arguments{
\item{data}{Benthic field data set.}

\item{dbase_type}{Production database to use, either Indo-Pacific ReefBudget ("IPRB")
or U.S. Pacific Islands NCRMP-specific database ("NCRMP"). The Indo-Pacific ReefBudget
database is derived from "IP Calcification and bioerosion rates database v.1.3",
downloaded from https://geography.exeter.ac.uk/reefbudget/indopacific/. Defaults to "NCRMP".}

\item{qc_check}{Remove transects where rugosity < 1.}

\item{sites_metadata}{Data frame containing information about sites}

\item{method_name}{Transect design by which data were collected ("IPRB", "Chords", or "SfM").}
}
\description{
Calculate carbonate production rates from benthic census data
}
\examples{
benthic_data <- read.csv("ESD_CarbBudget_Benthic_OAHU_2021.csv",
   na = "", check.names = FALSE)

prod_iprb <- process_prod(
   data = benthic_data,
)
}
\author{
Hannah Barkley
}
