% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_urchins.R
\name{process_urchins}
\alias{process_urchins}
\title{Calculate urchin erosion rates from urchin census data}
\usage{
process_urchins(
  data,
  transect_id = c("A1", "A2", "A3", "B1", "B2", "B3"),
  transect_length = c(10, 10, 10, 10, 10, 10),
  method_name = c("IPRB", "Chords"),
  data_type = c("In water", "SfM"),
  full_summary = FALSE
)
}
\arguments{
\item{data}{Urchin observation data set.}

\item{transect_id}{String of transect names (e.g., ("A1", "A2", "A3", "B1", "B2", "B3")).}

\item{transect_length}{String of transect lengths in meters (e.g., c(10, 10, 10, 10, 10, 10)).}

\item{method_name}{Transect design by which data were collected ("IPRB" or "Chords").}

\item{data_type}{Type of data collection ("In water" or "SfM").}
}
\description{
Calculate urchin erosion rates from urchin census data
}
\author{
Hannah Barkley
}
