\name{reefbudgetR-package}
\alias{reefbudgetR-package}
\alias{reefbudgetR}
\docType{package}

\title{reefbudgetR: Process ReefBudget carbonate budget data}

\description{ The \pkg{reefbudgetR} package tools for working with ReefBudget carbonate budget data, with functions to process field-based and SfM-derived benthic, urchin, and parrotfish census data and calculate carbonate production and erosion.  }

\details{The functions in the \pkg{reefbudgetR} package contain tools for carbonate budget analysis.
}


