% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_calc_prod.R
\name{run_calc_prod}
\alias{run_calc_prod}
\title{Run calc_prod function}
\usage{
run_calc_prod(data, transect_id, transect_length, dbase_type, data_type, ...)
}
\arguments{
\item{data}{Benthic data set to process}

\item{transect_id}{String of transect names}

\item{transect_length}{String of transect lengths in meters}

\item{dbase_type}{Production database to use ("CPM" or "NCRMP")}

\item{data_type}{Type of data collection ("In water" or "SfM")}
}
\description{
Run calc_prod function
}
\author{
Hannah Barkley
}
