% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_calc_prod.R
\name{run_calc_prod}
\alias{run_calc_prod}
\title{Run calc_prod function}
\usage{
run_calc_prod(data, dbase_type, prod_dbase_custom = NULL, ...)
}
\arguments{
\item{data}{Benthic data set to process}

\item{dbase_type}{Production database to use, either Indo-Pacific ReefBudget ("IPRB"), the
U.S. Pacific Islands NCRMP-specific database ("NCRMP"), or upload a customized database with location-specific rates ("Custom").
The Indo-Pacific ReefBudget database is derived from "IP Calcification and bioerosion rates database v.1.3",
downloaded from https://geography.exeter.ac.uk/reefbudget/indopacific/.}
}
\description{
Run calc_prod function
}
\examples{
benthic_data <- read.csv("ESD_CarbBudget_Benthic_OAHU_2021.csv",
    na = "", check.names = FALSE)

calc_prod_output <- run_calc_prod(
    data = benthic_data,
    dbase_type = "NCRMP"
    )
}
\author{
Hannah Barkley
}
