% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fish_erosion.R
\name{summarize_fish_erosion}
\alias{summarize_fish_erosion}
\title{Summarize site-level parrotfish bioerosion rates}
\usage{
summarize_fish_erosion(species_table, full_summary = FALSE)
}
\arguments{
\item{species_table}{Species table with average density, biomass, and bioersion metrics,
outputs of \code{summarize_fish_metrics}.}

\item{full_summary}{Return full summary of results (default is FALSE)}
}
\description{
Summarize site-level parrotfish bioerosion rates
}
\author{
Rebecca Weible
}
