% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_transect.R
\name{summarize_transect}
\alias{summarize_transect}
\title{Summarize benthic transect metadata}
\usage{
summarize_transect(
  data,
  transect_id,
  transect_length,
  data_type = c("In water", "SfM")
)
}
\arguments{
\item{data}{Production data set.}

\item{transect_id}{String of transect names (e.g., ("A1", "A2", "A3", "B1", "B2", "B3")).}

\item{transect_length}{String of transect lengths in meters (e.g., c(10, 10, 10, 10, 10, 10)).}

\item{data_type}{Type of data collection ("In water" or "SfM").}
}
\description{
Summarize benthic transect metadata
}
\author{
Hannah Barkley
}
