% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_transect.R
\name{summarize_transect}
\alias{summarize_transect}
\title{Summarize transect metadata}
\usage{
summarize_transect(
  data,
  transect_id,
  transect_length,
  data_type = c("In water", "SfM")
)
}
\arguments{
\item{data}{Production data set}

\item{transect_id}{String of transect names}

\item{transect_length}{String of transect lengths in meters}

\item{data_type}{Data type Type of data collection ("In water" or "SfM")}
}
\description{
Summarize transect metadata
}
\author{
Hannah Barkley
}
